requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
mysql-connector-python==8.0.26
